refactor: 불필요한 커스텀 훅 제거
